refactor: Header, NotiModal 대문자로 변경
