+++ frontend/src/components/Header/NotiModal/NotiModal.styled.ts
